bump travis JVM to JDK8
closure-template's pom.xml requires JDK8 (see the "1.8" in it). This
fixes the build by bumping the JVM TravisCI uses to JDK8.

Unfortunately, OpenJDK 8 is not supported on TravisCI yet, so we have to
switch to the Oracle JDK. Further, the version TravisCI has when you
just put "oraclejdk8" in the "jdk:" setting is a old buggy version that
crashes when compiling closure-templates.

Their support team told us to do this, instead.